requirements.txt
google-images-download==2.8.0
Pillow==8.0.1
psycopg2==2.8.6
selenium==3.141.0
urllib3==1.26.2
